HYDRAN-1869: Tweak api and entity

+++ src/main/java/se/sundsvall/memories/integration/db/FilmRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 import se.sundsvall.memories.integration.db.model.FilmEntity;
 
 @CircuitBreaker(name = "filmRepository")
-public interface FilmRepository extends JpaRepository<FilmEntity, Long>, JpaSpecificationExecutor<FilmEntity> {
+public interface FilmRepository extends JpaRepository<FilmEntity, Integer>, JpaSpecificationExecutor<FilmEntity> {
 
 }
